#30: clubInviteBuilder refactor - inlined invite construction to single place that used it - renamed function to clubInviteBuilder - moved clubInviteBuilder to messageEngine
