+++ source/commands/club-config.js
@@ -1,7 +1,8 @@
 const { ActionRowBuilder, ButtonBuilder, ButtonStyle } = require('discord.js');
 const Command = require('../classes/Command.js');
-const { getClubDictionary, clubInviteBuilder } = require("../helpers.js");
+const { getClubDictionary } = require("../helpers.js");
 const { SAFE_DELIMITER } = require('../constants.js');
+const { clubEmbedBuilder } = require('../engines/messageEngine.js');
 
 const options = [];
 const subcommands = [];
@@ -12,17 +13,19 @@ module.exports = new Command("club-config", "Change the configuration of the cur
  */
 module.exports.execute = (interaction) => {
 	const club = getClubDictionary()[interaction.channelId];
-	const { embeds } = clubInviteBuilder(club);
-	const components = [new ActionRowBuilder().addComponents(
-		new ButtonBuilder().setCustomId(`changeclubinfo${SAFE_DELIMITER}${club.id}`)
-			.setLabel("Set Name/Description")
-			.setStyle(ButtonStyle.Primary),
-		new ButtonBuilder().setCustomId(`changeclubseats${SAFE_DELIMITER}${club.id}`)
-			.setLabel("Set Members")
-			.setStyle(ButtonStyle.Primary),
-		new ButtonBuilder().setCustomId(`changeclubmeeting${SAFE_DELIMITER}${club.id}`)
-			.setLabel("Set Meeting Time")
-			.setStyle(ButtonStyle.Primary)
-	)];
-	interaction.reply({ embeds, components, ephemeral: true });
+	interaction.reply({
+		embeds: [clubEmbedBuilder(club)],
+		components: [new ActionRowBuilder().addComponents(
+			new ButtonBuilder().setCustomId(`changeclubinfo${SAFE_DELIMITER}${club.id}`)
+				.setLabel("Set Name/Description")
+				.setStyle(ButtonStyle.Primary),
+			new ButtonBuilder().setCustomId(`changeclubseats${SAFE_DELIMITER}${club.id}`)
+				.setLabel("Set Members")
+				.setStyle(ButtonStyle.Primary),
+			new ButtonBuilder().setCustomId(`changeclubmeeting${SAFE_DELIMITER}${club.id}`)
+				.setLabel("Set Meeting Time")
+				.setStyle(ButtonStyle.Primary)
+		)],
+		ephemeral: true
+	});
 }
